Preliminary ContentEncoder support
I’m not thrilled with the comprehensiveness of this — I think it should probably work for all functions down the encode stack, not just the top two as it currently does. Also would probably be nice to have some helper EncoderUnwrap function or something that would take an optional encoder and return the ContentEncoder (as a JSONEncoder) if the given encoder was nil.